// frontend/hub/interfaces/generated/SigningServiceResponse.ts
// Autogenerated file. Please do not modify.

// If you want to modify fields to interface, create new one in the folder above and create interface with the same name.
// You can then add, modify or even delete existing interface fields. Delete is done with Omit, note however it returns new interface.
// Those autogenerated interfaces does not contains all types, some of them are unknown - those are candidates for custom modification.
// See readme in folder above for more information.

/* eslint-disable @typescript-eslint/no-empty-interface */

// URL of interface:
// /api/pulp/api/v3/signing-services/{pulp_id}/

// A serializer for the model declaring a signing service.
export interface SigningServiceResponse {
  pulp_href: string;

  // Timestamp of creation.
  pulp_created: string;

  // A unique name used to recognize a script.
  name: string;

  // The value of a public key used for the repository verification.
  public_key: string;

  // The fingerprint of the public key.
  pubkey_fingerprint: string;

  // An absolute path to a script which is going to be used for the signing.
  script: string;
}